[YES-50] Refactor phase check, handle timeout case
getPhase() is now responsible for checking what phase the game is in so
the previous time checking code is no longer duplicated.
tick() disables an answer from being selected if the player did
not choose one in time since the disabling needs to be done immediately
after the answer selection phase ends, and it also clears the timer
as soon as it's done being used.