[YES-50] Incorporate new ACS into results screen
Also move the check for end of questions in TriviaGame to a lifecycle
method to keep render working as intended. Render should not change
state and/or props while it's rendering the component with the current
state and props which React warned me about, so since I wanted to check
after the component updated anyways, componentDidUpdate() makes a lot
more sense.